Persist the main editor's state (visual vs. text) regardless of FM
Without this, changing a FM editor's state from Viisual to Text (tinymce to html) would change the cookie value that WordPress stores for the main editor. Meaning that if a FM editor is changed from Visual to Text, the next time the main editor loads it will also be in Text mode.