Replace per-field JS handlers with Customizer event
This removes the two event handlers that existed only to enable sortable
fields and label macros. Instead, we trigger one event when a Customizer
section containing a Fieldmanager control expands, and have those
features listen for it. This is a little more consistent with other FM
features that listen for FM-related events.