Bind FM event to Sections as they're added
This helps ensure we bind 'fm_customizer_control_section_expanded' to
sections that are dynamically added.

This also changes the previous behavior of binding to only sections with
Fieldmanager controls: If an FM control is dynamically added to a
section that had no FM controls, we need to trigger the event when that
section expands.

Binding to when controls are added, rather than sections, isn't enough
because controls can be added without being assigned to sections.
Eventually, it might be more efficient to try binding to a section only
when a Fieldmanager control is added to it.