Serialize FM Control values into JS objects
When previewing or saving a Fieldmanager field in the Customizer, we now
use the jquery.serializeJSON][1] library to serialize the FM control
into a JavaScript object. The object is then set as value of the
control's setting.

This allows us to send an object both to PHP, as part of a 'refresh'
setting, and to JS, as part of a 'postMessage' setting. Previously each
received a string from jQuery's `serialize()`, which remains the
fallback behavior.

[1]: https://github.com/marioizquierdo/jquery.serializeJSON