Pass `$append = true` after first save when datasource repeats
When a Term datasource is a part of a repeatable group, the expected
behavior is to save all terms present in all groups, not just the term
of the last group. By passing `$append = true` after the first save, we
can ensure subsequent terms maintain their relationship to the post.