Add network submenu context
This adds `Fieldmanager_Context_NetworkSubmenu`, which extends
`Fieldmanager_Context_Submenu` with some network-friendly variations.

Network submenus can be activated during the new 'network_submenu'
context action, which helps Fieldmanager understand whether developers
intend to register a "regular" or network submenu.

The submenu context's default `$capability` is now set on the property
itself, not in the constructor, to make it easier to override.
`fm_register_submenu_page()` and `Fieldmanager_Field::add_submenu_page()`
defer to the contexts to determine their default capability.

The hook into 'admin_menu' for registering a submenu now happens in a
protected method, also for overriding.

The submenu context now also exits the constructor if no field was
passed to it. Although the default value for `$fm` is null, the
constructor assumes it's a field. Exiting also allows us to access
properties on an "inert" field, which `fm_register_submenu_page()` uses.