Reject invalid values from sanitize callback
When the context can use the Customizer's validation framework, and when
the default callbacks are added to FM Customizer settings, any invalid
submissions will have already been rejected before sanitizing occurs.

But, we still need to check for invalid values in the sanitizing
callback just in case, for example, the method is called on its own or
the validation callback is not hooked to the Customizer setting. If a
field value is invalid, we need to reject the update with a
Customizer-approved `WP_Error` or `null` before Fieldmanager's native
validation routines (such as throwing an Exception or calling
`wp_die()`) kick in.