minimum_count: Exactly what it says on the tin
This commit will cause the remove handles on repeated groups to not work
if the count of the repeated group is less than or equal to the
minimum_count. This brings the minimum_count feature closer to working
as I believe it was intended, and less of a clone of features already
offered by starting_count and extra_elements.