Finished updating unit tests
My revised code works by taking inspiration from the code for the "Add
Field" button. Each remove multi-tool now gets the data-minimum_count
attribute if such is set on the element. This caused it to fail unit
tests that were not prepped with the optional attribute. Now I have
edited the unit tests to assert that those that test the minimum_count
assert with preg_match_all instead of substr_count in the first
assertion for each function.