Cast term_taxonomy_id to string before pushing it to post_meta / update_option
Attempting to store integer in post_meta and option leads skipping false-negative result of new and old value comparison, which results in an extra SELECT and UPDATE SQL query, when the UPDATE fails anyway. More details on the issue can be found here: https://core.trac.wordpress.org/ticket/22192

This commit is casting the integer value of term_id / term_taxonomy_id to string before it's passed to update_post_meta function. This fixes the false-negative result of the mentioned comparison and thus saves 2 SQL queries in case the value does not get changed.

The change introduced here should be compatible with any future change to the update_option and update_metadata functions, at least as proposed in patches in mentioned ticket.