Move the `[]` appended to form names to the base `Options` class
The `[]` had been appended manually to `Fieldmanager_Select` and
`Fieldmanager_Checkboxes` form names. But, `Fieldmanager_Util_Validation`
relies only on `Fieldmanager_Field::get_form_name()` for determining
the name, which wouldn't include `[]`.

With this change, the `[]` in multi-selects and checkboxes is
generated by overriding `Fieldmanager_Field::get_form_name()` and
checking the `$multiple` property, which ensures `[]` is also returned
to other callers like `Fieldmanager_Util_Validation`.

However, this change assumes that any field with `'multiple' => true`
should in fact be represented as an array in its form name, which is
true for the bundled fields but might not be true for custom
classes. Those custom classes would now need to override
`get_form_name()` to ensure it behaved as expected.