Deprecate `fm_url_to_post_id()`
Support for custom post types was added to `url_to_postid()` in WordPress
3.7. According to its own docs and the site where it came from,
`fm_url_to_post_id()` was intended to work around the previous lack of
support for CPTs.

Also corrects the return type for `fm_url_to_post_id()`, which is
important in this case for an assurance of compatibility with
`url_to_postid()`.